pyproject.toml
[project]
name = "smart-task-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf2>=2.8.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
]
